Add Sprint Stamina Cost
CharacterAttributeEntity CurrentValue is now a float. (Database)
CharacterResourceAttribute CurrentValue is now a float.
UIResourceBar will now round to nearest Int.
Default sprint stamina cost is 5pts/second.

+++ FishMMO-Database/FishMMO-DB/Npgsql/Entities/Scene/Character/CharacterAttributeEntity.cs
@@ -13,6 +13,6 @@ namespace FishMMO.Database.Npgsql.Entities
         public CharacterEntity Character { get; set; }
         public int TemplateID { get; set; }
         public int Value { get; set; }
-        public int CurrentValue { get; set; }
+        public float CurrentValue { get; set; }
     }
 }
